Reorganized the project file layout. This was a long time in coming, especially since it hadn't been done since WM became a node-based system. (!)
Misc modules have been moved out into "More Patches." This provides a nice location to organize random stuff we decide to add to the project and perhaps eventually make into a separate repo with contributions from users of WM.

Ports are all collected together.

Render is more coherent.

Hopefully, the structure of the code is easier to understand now looking at the folder layout.

+++ Engine/Ports/WMNumberPort.h
@@ -10,10 +10,9 @@
 
 #import "WMPort.h"
 
-//TODO: offer a port with double or float values for performance?
 @interface WMNumberPort : WMPort {
 }
-@property (nonatomic) double value;
+@property (nonatomic) float value;
 
 
 @end
